[connectivity_for_web] Rename plugin after rebasing.
