+++ packages/connectivity/connectivity_for_web/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.0
+
+* Rename from "experimental_connectivity_web" to "connectivity_for_web", and move to flutter/plugins master.
+
 ## 0.2.0
 
 * Add fallback on dart:html for browsers where NetworkInformationAPI is not supported.
