+++ packages/connectivity/connectivity_for_web/ios/connectivity_for_web.podspec
@@ -3,13 +3,13 @@
 # Run `pod lib lint connectivity_web.podspec' to validate before publishing.
 #
 Pod::Spec.new do |s|
-  s.name             = 'connectivity_web'
+  s.name             = 'connectivity_for_web'
   s.version          = '0.1.0'
-  s.summary          = 'No-op implementation of connectivity_web web plugin to avoid build issues on iOS'
+  s.summary          = 'No-op implementation of connectivity web plugin to avoid build issues on iOS'
   s.description      = <<-DESC
 temp fake connectivity_web plugin
                        DESC
-  s.homepage         = 'https://github.com/flutter/plugins/tree/master/packages/connectivity/connectivity_web'
+  s.homepage         = 'https://github.com/flutter/plugins/tree/master/packages/connectivity/connectivity_for_web'
   s.license          = { :file => '../LICENSE' }
   s.author           = { 'Flutter Team' => 'flutter-dev@googlegroups.com' }
   s.source           = { :path => '.' }
