+++ packages/connectivity/connectivity_for_web/lib/src/utils/connectivity_result.dart
