+++ packages/connectivity/connectivity_for_web/pubspec.yaml
@@ -1,6 +1,6 @@
-name: experimental_connectivity_web
-description: An (experimental) implementation for the web platform of the Flutter `connectivity` plugin. This uses the NetworkInformation Web API, with a fallback to Navigator.onLine.
-version: 0.2.0
+name: connectivity_for_web
+description: An implementation for the web platform of the Flutter `connectivity` plugin. This uses the NetworkInformation Web API, with a fallback to Navigator.onLine.
+version: 0.3.0
 homepage: https://github.com/ditman/plugins/tree/connectivity-web/packages/connectivity/experimental_connectivity_web
 
 flutter:
@@ -8,7 +8,7 @@ flutter:
     platforms:
       web:
         pluginClass: ConnectivityPlugin
-        fileName: experimental_connectivity_web.dart
+        fileName: connectivity_for_web.dart
 
 dependencies:
   connectivity_platform_interface: ^1.0.3
