--- packages/connectivity/experimental_connectivity_web/android/gradle.properties
@@ -1,4 +1,0 @@
-org.gradle.jvmargs=-Xmx1536M
-android.enableR8=true
-android.useAndroidX=true
-android.enableJetifier=true
