--- packages/connectivity/experimental_connectivity_web/android/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'connectivity_web'
