--- packages/connectivity/experimental_connectivity_web/android/src/main/AndroidManifest.xml
@@ -1,3 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.example.connectivity_web">
-</manifest>
